duplicateDetection/requirements.txt
pandas
fastapi
uvicorn
scikit-learn
numpy
pdfplumber
python-docx
python-multipart
joblib
torch
transformers
openpyxl
tensorflow
accelerate
xgboost
torch 
torchvision 
ftfy 
regex 
tqdm
fitz
frontend
tools
PyMuPDF
pytesseract